requirements.txt
wheel
requests>=2.28.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
sentence-transformers>=2.2.2
datasets>=2.12.0
rouge>=1.0.1
pinecone
boto3==1.35.88
opensearch-py==2.8.0
nltk
accelerate
plyvel
# farm-haystack
uvicorn
streamlit
fastapi
psycopg2-binary>=2.9.0
llama-index
llama-index-retrievers-bm25

# language detection
langdetect
